refactor: reorganize the route models
